chore(warnings): progress commit - clear C4244 + C4996 categories
Adds explicit static_cast<DestType>(...) on lossy numeric conversions (131 unique sites / 187 raw across ~30 files), retypes a few loop counters to collapse clusters, and wraps std::transform(s, ::toupper/::tolower) calls in [](unsigned char) -> char lambdas to silence the int -> char narrowing in <algorithm>. Replaces the 4 remaining strncpy sites with equivalent-semantics memcpy and the 5 getenv sites with a new safe_getenv() -> std::string helper using _dupenv_s on MSVC. C4244 and C4996 both at zero; all 16 categories from this audit are now cleared, first-party warnings = 0. No public API changes; clean rebuild reports 0 errors / 0 warnings, all 6 ctest suites pass. Updates WARNINGS.md to reflect live state.

+++ src/controller/src/compute/rocprofvis_controller_trace_compute.cpp
@@ -500,7 +500,7 @@ rocprofvis_result_t ComputeTrace::LoadRocpd(Future* future)
                                                 }
                                                 workload->SetObject(kRPVControllerWorkloadKernelIndexed, i, (rocprofvis_handle_t*)kernel);
                                                 uint64_t id = std::stoull(data_store.rows[i][data_store.columns.at(kRPVComputeColumnKernelUUID).value()]);
-                                                kernel_ids.push_back(id);
+                                                kernel_ids.push_back(static_cast<uint32_t>(id));
                                             }
                                         });
                                     }                                   
@@ -511,7 +511,7 @@ rocprofvis_result_t ComputeTrace::LoadRocpd(Future* future)
                                         std::optional<double> min_intensity_x;
                                         std::optional<double> max_intensity_y;
                                         std::optional<double> min_intensity_y;
-                                        double uint_data = 0;
+                                        uint64_t uint_data = 0;
                                         for(const uint32_t& kernel_id : kernel_ids)
                                         {
                                             m_query_arguments = { {kRPVComputeParamKernelId, std::to_string(kernel_id)} };
